pass1: memory diet - dead fields, per-function frees, packed names
Three heap reductions plus the pass1 side of enum lowering, all
verified byte-identical through the native/Z80 equivalence matrix
(now 27/27 SAME; make footprint):

- struct expr 26 -> 20 bytes: stmt and inst had zero uses anywhere in
  pass1; prev had one write and no reads (args walk via next only).

- parsefunc frees each function's scope names as soon as the function
  completes instead of parking them on the graveyard until exit: the
  graveyard is LIFO, so everything above a snapshot taken at function
  entry is ours.  Phase 1 already captured u.locals copies; those are
  freed too once phase 2 finishes emitting the function.  Name
  high-water drops from every-name-in-the-file to file scope plus the
  current function.

- struct name 39 -> 35 bytes: the member group (offset/bitoff/width)
  and the regalloc group (ref_count/agg_refs/reg/addr_taken/frm_off)
  overlay in a union - no name is both a struct member and a register
  candidate - and static_id narrows to uchar to match its counters.
  zc3 enums are already 1 byte, so kind needed no change.

- enum machinery deleted: cpp's filtenum guarantees ENUM never
  arrives, so getbasetype's enum block and the ENUM statement cases
  go away (-245 bytes of text).  Prerequisite: the kind constants are
  renamed kprim/kvar/kelem/... - the old names collided with real
  members and declarators (struct var *var), which the filter's
  textual expansion cannot scope around.

- expr.c constant ladder: spell the sval >= 0 branch as plain longtype
  instead of sval <= 2147483647L, which zc3 rewrites to < LONG_MIN
  (always false), mistyping LONG_MAX constants as unsigned long.

Worst-case c0 heap gap on the self-compile corpus went from 2088
bytes to 6065+ across these changes.  Baselines re-blessed for the
full corpus, including the pending .err-file retirement and rc scheme
from the driver work.

+++ ccc/pass1/util.c
@@ -294,7 +294,9 @@ void asmDbStr(unsigned char *data, int len)
 void asmDw(int val)
 {
 	char buf[16];
-	fmtstr(buf, "\t.dw %d", val & 0xffff);
+	/* mask in long so 16-bit and 32-bit int hosts print the same
+	 * string for negative values (65535, not -1) */
+	fmtstr(buf, "\t.dw %ld", (long)val & 0xffff);
 	asmLine(buf);
 }
